# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * DigitalXNews API - AI-generated Islamic news platform
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Post {
  id: string;
  title: string;
  body: string;
  category: string;
  /** @nullable */
  imageUrl?: string | null;
  hasImage: boolean;
  significanceScore: number;
  sourceNote: string;
  publishedAt: string;
  expiresAt: string;
  isBreaking: boolean;
  createdAt: string;
  /** @nullable */
  titleEn?: string | null;
  /** @nullable */
  bodyEn?: string | null;
  /** @nullable */
  titleUr?: string | null;
  /** @nullable */
  bodyUr?: string | null;
  /** @nullable */
  titleAr?: string | null;
  /** @nullable */
  bodyAr?: string | null;
  likesCount?: number;
  viewsCount?: number;
}

export interface EngagementResponse {
  likesCount: number;
  viewsCount?: number;
}

export interface PostListResponse {
  posts: Post[];
  total: number;
  page: number;
  limit: number;
}

export interface CategoryCount {
  category: string;
  count: number;
}

export interface UserPreference {
  id: string;
  deviceId: string;
  /** @nullable */
  pushToken: string | null;
  followedCategories: string[];
  notificationsEnabled: boolean;
  createdAt: string;
}

export interface PreferenceInput {
  deviceId: string;
  /** @nullable */
  pushToken?: string | null;
  followedCategories?: string[];
  notificationsEnabled?: boolean;
}

export interface FlaggedPost {
  id: string;
  title: string;
  body: string;
  category: string;
  significanceScore: number;
  sourceNote?: string;
  flagReason: string;
  flaggedAt: string;
  createdAt: string;
}

export interface ErrorResponse {
  error: string;
}

export type ListPostsParams = {
/**
 * Filter by category name
 */
category?: string;
page?: number;
limit?: number;
};