# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DigitalXNews API - AI-generated Islamic news platform
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: posts
    description: News post operations
  - name: preferences
    description: User preferences and push notification registration
  - name: admin
    description: Content moderation operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /posts:
    get:
      operationId: listPosts
      tags: [posts]
      summary: List news posts
      description: Returns paginated news posts, optionally filtered by category
      parameters:
        - name: category
          in: query
          schema:
            type: string
          description: Filter by category name
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostListResponse"

  /posts/categories:
    get:
      operationId: listCategories
      tags: [posts]
      summary: List available categories
      description: Returns all categories with post counts
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategoryCount"

  /posts/{id}:
    get:
      operationId: getPost
      tags: [posts]
      summary: Get a single post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Post detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "404":
          description: Post not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /posts/{id}/like:
    post:
      operationId: likePost
      tags: [posts]
      summary: Increment likes count for a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated likes count
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EngagementResponse"
        "404":
          description: Post not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /posts/{id}/view:
    post:
      operationId: viewPost
      tags: [posts]
      summary: Increment view count for a post (call once per device per post)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated view count
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EngagementResponse"
        "404":
          description: Post not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /preferences:
    post:
      operationId: upsertPreferences
      tags: [preferences]
      summary: Register or update device push preferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PreferenceInput"
      responses:
        "200":
          description: Preferences saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPreference"

  /preferences/{deviceId}:
    get:
      operationId: getPreferences
      tags: [preferences]
      summary: Get device preferences
      parameters:
        - name: deviceId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Device preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPreference"
        "404":
          description: Preferences not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/flagged:
    get:
      operationId: listFlaggedPosts
      tags: [admin]
      summary: List flagged posts awaiting moderation
      responses:
        "200":
          description: List of flagged posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FlaggedPost"

  /admin/flagged/{id}/approve:
    post:
      operationId: approvePost
      tags: [admin]
      summary: Approve a flagged post and publish it
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Post approved and published
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "404":
          description: Flagged post not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/flagged/{id}/reject:
    post:
      operationId: rejectPost
      tags: [admin]
      summary: Reject and delete a flagged post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Post rejected and deleted
        "404":
          description: Flagged post not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Post:
      type: object
      required: [id, title, body, category, hasImage, significanceScore, sourceNote, publishedAt, expiresAt, isBreaking, createdAt]
      properties:
        id:
          type: string
        title:
          type: string
        body:
          type: string
        category:
          type: string
        imageUrl:
          type: ["string", "null"]
        hasImage:
          type: boolean
        significanceScore:
          type: integer
        sourceNote:
          type: string
        publishedAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        isBreaking:
          type: boolean
        createdAt:
          type: string
          format: date-time
        titleEn:
          type: ["string", "null"]
        bodyEn:
          type: ["string", "null"]
        titleUr:
          type: ["string", "null"]
        bodyUr:
          type: ["string", "null"]
        titleAr:
          type: ["string", "null"]
        bodyAr:
          type: ["string", "null"]
        likesCount:
          type: integer
          default: 0
        viewsCount:
          type: integer
          default: 0

    EngagementResponse:
      type: object
      required: [likesCount]
      properties:
        likesCount:
          type: integer
        viewsCount:
          type: integer

    PostListResponse:
      type: object
      required: [posts, total, page, limit]
      properties:
        posts:
          type: array
          items:
            $ref: "#/components/schemas/Post"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    CategoryCount:
      type: object
      required: [category, count]
      properties:
        category:
          type: string
        count:
          type: integer

    UserPreference:
      type: object
      required: [id, deviceId, pushToken, followedCategories, notificationsEnabled, createdAt]
      properties:
        id:
          type: string
        deviceId:
          type: string
        pushToken:
          type: ["string", "null"]
        followedCategories:
          type: array
          items:
            type: string
        notificationsEnabled:
          type: boolean
        createdAt:
          type: string
          format: date-time

    PreferenceInput:
      type: object
      required: [deviceId]
      properties:
        deviceId:
          type: string
        pushToken:
          type: ["string", "null"]
        followedCategories:
          type: array
          items:
            type: string
        notificationsEnabled:
          type: boolean

    FlaggedPost:
      type: object
      required: [id, title, body, category, significanceScore, flagReason, flaggedAt, createdAt]
      properties:
        id:
          type: string
        title:
          type: string
        body:
          type: string
        category:
          type: string
        significanceScore:
          type: integer
        sourceNote:
          type: string
        flagReason:
          type: string
        flaggedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string