# lib/api-zod/src/generated/types/categoryCount.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * DigitalXNews API - AI-generated Islamic news platform
 * OpenAPI spec version: 0.1.0
 */

export interface CategoryCount {
  category: string;
  count: number;
}