# lib/api-zod/src/generated/types/engagementResponse.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * DigitalXNews API - AI-generated Islamic news platform
 * OpenAPI spec version: 0.1.0
 */

export interface EngagementResponse {
  likesCount: number;
  viewsCount?: number;
}