# lib/api-zod/src/generated/types/flaggedPost.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * DigitalXNews API - AI-generated Islamic news platform
 * OpenAPI spec version: 0.1.0
 */

export interface FlaggedPost {
  id: string;
  title: string;
  body: string;
  category: string;
  significanceScore: number;
  sourceNote?: string;
  flagReason: string;
  flaggedAt: Date;
  createdAt: Date;
}