# lib/api-zod/src/generated/types/postListResponse.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * DigitalXNews API - AI-generated Islamic news platform
 * OpenAPI spec version: 0.1.0
 */
import type { Post } from './post';

export interface PostListResponse {
  posts: Post[];
  total: number;
  page: number;
  limit: number;
}